Force (re)encoding of UTF-8 in `Burlap.parser`
A Burlap API is incorrectly escaping emoji as a series of hex bytes that
are illegal UTF-8 byte sequences.

The HTTP response is not reporting its character set so is interpreted
as UTF-8 but the illegal sequence causes Nokogiri to immediate abort
returning `nil` which goes on to cause worse problems.

Using `encode` on the parser input does not change UTF-8 strings but
will safely replace the hex digits with valid data. This allows the
parsing to continue.

There is an optional second parameter to pass the encoding to use if it
is available, otherwise UTF-8 will be used.