Sanitise ASCII before `Burlap.parse` handoff
Missed the more common case where the API response, without encoding is
actually being encoded as ASCII_8BIT. The previous fix was able to raise
an encoding error when encountering bad UTF-8 bytes but that error did
not help the real issue. Just prevented pushing it down to silently
return in Nokogiri.

This uses a regexp taken from the following article that sanitises the
hex values and clears them out.

See https://www.johnhawthorn.com/2009/10/sanitizing-utf8-in-ruby/

Again, not brilliant but a sanitised output is preferable to the
incorrect behaviour when the rest of the Burlap XML is discarded due to
the text contents of a few nodes.