Remove off spec Burlap workarounds
The last few commits were work arounds for a very poorly implemented API
causing issues. Neither helped the real issues.

Going back to basic and the Burlap spec includes:

> Burlap uses SML (Simple Markup Language), a restricted subset of XML:
>
> ...
> * Only UTF-8 encoding is allowed.
> * Only numbered character entities, e.g. &#60;, and the standard XML entities (&lt;, &gt;, and &amp;) are allowed as escapes.
> ...

http://hessian.caucho.com/doc/burlap-1.0-spec.xtp

So the API returning ASCII encoded XML with incorrectly escaped emoji is
not really a problem burlap should be solving.

This removes the encoding nonsense and adds a `strict` option which,
when `true` will attempt to force the encoding of the string to UTF-8.

This may raise an `Encoding` error if the input is garbage or can handle
it correctly as before.

The `strict = false` route is the original code where the error will
cause Nokogiri to silently return `nil` as the code did previously.