Remove `--deployment` to not dictate gems
Since `burlap` is a library, the dependencies should be defined in the
Gemspec and managed by the application/bundler.

Including a `--deployment` argument to bundler prevents installing gems
without the `Gemfile.lock` we removed because it should not have been
committed in a library.

Different Ruby versions may pull in different dependencies and locking
the expected version for testing is not testing them correctly for out
of date versions or issues which need resolving in the Gemspec.