Replace Proc "hack" to work with Ruby 3.0+
The `DefaultResolver` relied on a self-declared "hack" which was later
deprecated and removed in Ruby 3.0. It relied on a check for a given
block and then assignment of a `Proc`.

The fix involves explicitly including the `&block` argument and removing
the check for it, since the check fails on earlier ruby versions but the
code itself works.

The spec that checked for the `ArgumentError` raised in this case was
thus broken and so removed.